# seance 7 & 8 - Ionic/openapi: '3.0.2'.yml
openapi: '3.0.2'
info:
  title: Tragency
  description: Seance 6 of the course web architecture
  version: '1.0'
servers:
  - url: http://localhost:3000
paths:
...
  /destinations/add:
    post:
      summary: Add a destination
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              $ref: '#/components/schemas/Destination'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/GoodResponse'
  /destinations/search?term={term}:
    get:
      summary: Search with a term and get all matching destinations
      parameters:
        - name: term
          in: path
          required: true
          description: Parameter description in CommonMark or HTML.
          schema:
            type: string
            minimum: 1
            example: Nice
      responses:
        '200':
          description: A JSON dict of destination
          content:
            application/json:
              schema:
                type: array
                $ref: '#/components/schemas/Destinations'

  # AGENCY
...
  /agencies/add:  
    post:
      summary: add new agency
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              $ref: '#/components/schemas/Agency'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/GoodResponse'
...
  /agencies/contact:
    post:
      summary: Send a form to contact the selected agency
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                destination_id:
                  type: integer
                  format: int64
                  example: 4
                  required:
                    - "true"
                agency_id:
                  type: integer
                  format: int64
                  example: 6
                  required:
                    - "true"
                name:
                    type: string
                    example: "Ivan Totravel"
                email:
                  type: string
                  example: ivan@gmail.com
                phone:
                  type: string
                  example: +324871937

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/GoodResponse'


  # LINKS
...
components:
  schemas:
    GoodResponse:
      description: A good response
      properties:
        message: 
          type: string
          example: Operation done
    Destination:
      description: Model containing Destination info
      properties:
        # id:
        #   type: integer
        #   example: 1
        name:
          type: string
          example: Ballito Green Estate
        description:
          type: string
          example: Big Estate
        country:
          type: string
          example: South Africa  
    Destinations:
      description: Model containing multiple destinations
      items:
        type: object         
        $ref: '#/components/schemas/Destination'
    Agency:
      description: Model contaning agency info
      properties:
        name:
          type: string
          example: WeKeepYourMoney
        email:
          type: string
          example: wekeepyourmoney@gmail.com
        phone:
          type: string
          example: +324871937
        localization:
          type: string
          example: wall street n°8
...